dream_store/requirements.txt
asgiref==3.7.2
Django==3.2.3
django-filter==23.3
djangorestframework==3.12.4
djoser==2.2.0
Pillow==10.0.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
